WIP: [infra] Update infer for OCaml 5
Summary:
This is an MVP to make Infer build with ocaml 5.0.0~alpha1

It does require a bit of a setup, in particular:
1. opam repo add alpha git+https://github.com/kit-ty-kate/opam-alpha-repository.git
2. `opam pin add` for [javalib](https://github.com/javalib-team/javalib/pull/22) and
   [sawja](https://github.com/javalib-team/sawja/pull/8).

parmap is not compatible with multi-core and I'm not sure it'll be
compatible anytime soon because of [compatibility.h
deprecation](https://github.com/ocaml/ocaml/pull/10863). However, we use
only `Setcore` from parmap; some of those functions should have an
alternative in Stdlib.Domain.